Full Bot For Sale/requirements.txt
telebot
telethon
pyrogram
asyncio
bs4
fake_useragent
googlesearch-python